requirements.txt
anyio==3.6.2
certifi==2023.5.7
fern-belvo==0.0.33
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
pydantic==1.10.7
rfc3986==1.5.0
sniffio==1.3.0
typing_extensions==4.5.0
